parts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
cachetools==5.5.0
    # via google-auth
certifi==2024.7.4
    # via requests
cffi==1.17.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==43.0.0
    # via pyjwt
django==5.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-storages
django-allauth[socialaccount]==64.1.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-storages[google]==1.14.4
    # via -r requirements.in
google-api-core[grpc]==2.19.2
    # via
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.34.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-secret-manager==2.20.2
    # via -r requirements.in
google-cloud-storage==2.18.2
    # via django-storages
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.65.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via google-cloud-secret-manager
grpcio==1.66.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.66.1
    # via google-api-core
idna==3.7
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==5.28.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via django-allauth
requests==2.32.3
    # via
    #   django-allauth
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
rsa==4.9
    # via google-auth
sqlparse==0.5.1
    # via django
urllib3==2.2.2
    # via requests
